==> src/00_import.R <==
# Author:NadineBestard
# Date 26/02/21

#Here I import the count matrices generated by CellRanger into R
#using Bioconductor packages. I will use the filtered matrices where an initial
#cellcalling has been done by CellRanger. In Cellranger version 7 the cellcalling
#is done with a very similar algorithm to dropletutilities one.

# symlincs have been created between main directory and data/cellranger

# Set up ------
library(here)
library(DropletUtils)

# Create file path for each sample matrix and metadata ------
# Select only the old samples
samples <- dir(here("data/CellRanger/"))
matrices <- here("data/CellRanger/", samples, "filtered_feature_bc_matrix")
metadata <- read.csv(here("data/metadata_scRNAseq.csv"))
metadata <- metadata[metadata$tissue == "thalamus", ]

# Create object -----
sce <- read10xCounts(matrices, samples, version = "auto", col.names = TRUE)
# add the metadata
coldata <- colData(sce)
# keep the rownames (same as colnames of sce)
names <- row.names(coldata)
coldata <- merge(x= coldata, y= metadata, by = "Sample", all.x=TRUE)
# reset the rownames to previous value, as there're lost with merge
row.names(coldata) <- names
# assign new coldata back to sce 
colData(sce) <- coldata
# The chip is detected as an integer instead of a character
sce$chip <- as.character(sce$chip)
# Save the object
saveRDS(sce, here("processed", "sce_th.RDS"))
