==> src/10_norm_featureselect_dimred_02.Rmd <==
---
title: "Reanalise after QC"
author: "Nadine Bestard"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: hide
editor_options: 
  markdown: 
    wrap: 72
---

# Set-up

```{r output-code, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r set-up, message=FALSE, warning=FALSE, eval=TRUE }
library(here) # for reproducible paths
library(SingleCellExperiment)
library(scater) # For qcs
library(scran) # for normalisation, feature selection
library(ggplot2) # To add titles to plots
library(patchwork) # agregate plots
library(pals) # for palettes with large n
```

```{r load-sce, eval=TRUE}

project <- "fire-mice"
sce <- readRDS(here("processed", project, "sce_clusterQC.RDS"))
source(here("src/colours.R"))

```

The object has `r dim(sce)[1]` genes and `r dim(sce)[2]` cells.

# Normalisation

```{r normalisation}
## Normalise by deconvolution ---
if(!file.exists(here("processed", project,  "sce_norm_02.RDS"))){
# For reproducibility
set.seed(100)
# Quick clustering to pool samples together and deal with 0 counts
quick_clusters <- quickCluster(sce)
# Calculate size factors
sce <-
  computeSumFactors(sce, cluster = quick_clusters, min.mean = 0.1)
# Check that there are not negative size factors
summary(sizeFactors(sce))
# Apply size factors and log transform them
sce <- logNormCounts(sce)
# save object
saveRDS(sce, here("processed", project,  "sce_norm_02.RDS"))
}
```

# Feature selection

After the quality control the variable genes need to be selected again
to better represent the variance in this cleaned dataset. We follow the
same methods than for our first feature selection, selecting here for
the top 2000 genes.

```{r featureselect}
gene_var_df <- modelGeneVar(sce)
gene_var <- metadata(gene_var_df)
plot(gene_var$mean, gene_var$var, xlab= "Mean of log-expression", ylab= "Variance of log-expression")
curve(gene_var$trend(x), lwd=2, add=T, col = "red")
# select hvgs
hvgs <- getTopHVGs(gene_var_df, n=2000)
# save them in the object
rowSubset(sce) <- hvgs
```

# Dimensionality reduction

## Run PCA and choose PCs

```{r }
if(!file.exists(here("processed", project,  "sce_dimred_02.RDS"))){
set.seed(1000)
sce <- runPCA(sce)

pct_var <- attr(reducedDim(sce, "PCA"), "percentVar")
plot(pct_var, log="y", xlab="PC", ylab="pct variance explained")



 #will keep 29PCs 

#keep the previous dimensional reduction just in case
reducedDim(sce, "PCA_all") <- reducedDim(sce, "PCA")
# And replace the default PCA with the reduced PCs
reducedDim(sce, "PCA") <- reducedDim(sce, "PCA")[,1:29]
}
```

## Visualisation

For visualisation, reduce to 2 dimensions. Non linear reductions.

### UMAP and TSNE

```{r eval=TRUE}
if(!file.exists(here("processed", project,  "sce_dimred_02.RDS"))){
set.seed(1000)
sce <- runUMAP(sce, dimred="PCA")
plotReducedDim(sce, dimred="UMAP", colour_by="genotype", point_size=0.1, point_alpha = 0.3)
set.seed(1000)
sce <- runTSNE(sce, dimred="PCA")
plotReducedDim(sce, dimred="TSNE", colour_by="genotype", point_size=0.1, point_alpha = 0.3)
plotReducedDim(sce, dimred="TSNE", colour_by="chip", point_size=0.1, point_alpha = 0.3)
plotReducedDim(sce, dimred="TSNE", colour_by="sex", point_size=0.1, point_alpha = 0.3)


saveRDS(sce, here("processed", project, "sce_dimred_02.RDS"))

}else{ # load teh corrected object
  sce <- readRDS(here("processed", project,  "sce_dimred_02.RDS"))
  plotReducedDim(sce, "TSNE", colour_by="sex")
}
```
