==> src/12_clustering_02.Rmd <==
---
title: "Clustering"
author: "NadineBestard"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: hide
editor_options: 
  markdown: 
    wrap: 72
    
---

### Set-up


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

The workflow and explanations bellow are from [OSCA](https://bioconductor.org/books/release/OSCA/dimensionality-reduction.html)

```{r set-up, message=FALSE, warning=FALSE}
library(SingleCellExperiment)
library(here) #reproducible paths
library(scater) # Plot dimred
library(bluster) # Clustering 
library(clustree) # show relationship clustering
library(Seurat) # clusternig 2
library(pals) # for palettes with large n #kelly()22, #polychrome()#36, unname(alphabet())

```

```{r}
project<- "fire-mice"
source(here("src/colours.R"))
```

```{r load-sce}
if(!file.exists(here("processed", project, "sce_clusters_02.RDS"))){
sce <- readRDS(here("processed", project, "sce_corrected_02.RDS"))
} else {
  sce <- readRDS(here("processed", project, "sce_clusters_02.RDS"))
}
```

```{r functions, echo=FALSE}
# plot all the seurat or sce dim reduction
plot_list_func <- function(obj,
                           col_pattern="RNA_snn_res.", 
                           plot_cols, 
                           clust_lab = TRUE,
                           label_size = 1,
                           num_col = 4,
                           reduction = "umap",
                           # todo add a save FALSE
                           save_dir = getwd(),
                           width=7,
                           height=5){
if(class(obj)[1] == "Seurat"){
  extr_res_col <- grep(pattern = col_pattern, names(obj@meta.data))
  res_names <- names(obj@meta.data[extr_res_col])
  # gtools function, sorts gene_names alphanumeric:
  res_names <- gtools::mixedsort(res_names) 
  for(i in 1: length(res_names)){
  #pdf(paste0(save_dir, 
    #           res_names[i], "_umap.pdf"), width=width, height=height)
  dim_plot <- DimPlot(obj, 
                          reduction = reduction, 
                          cols= plot_cols,
                          group.by = res_names[i],
                          label = clust_lab,
                          label.size = label_size) + NoLegend()
  print(dim_plot)
 # dev.off()
 # print(dim_plot)
  }
}else if (class(obj)[1] == "SingleCellExperiment"){
  names <- grep(pattern = col_pattern, names(colData(sce)), value = TRUE)
  ks <- stringr::str_remove(string = names, pattern = col_pattern)
  # gtools function, sorts gene_names alphanumeric:
  names <- gtools::mixedsort(names) 
  # plot directly from the sce_clusters_01.RDS
  idx <- c(1:length(names))
  for (i in idx) {
    k <- ks[i]
    cluster_name <- names[i]
    plotDim <-
      plotReducedDim(
        obj,
        dimred = reduction,
        colour_by = cluster_name,
        point_size = label_size,
        point_alpha = 0.3,
        text_by = cluster_name,
        text_size = 3
      )  + scale_colour_manual(values = plot_cols) + ggtitle(paste("Clustering with resolution/k =", k))
    print(plotDim)
  }
}else{
  stop("object must be seurat or singlecellexperiment")
}
  
}

```

## Motivation

Clustering is an unsupervised learning procedure that is used in scRNA-seq data analysis to empirically define groups of cells with similar expression profiles. It is worth stressing the distinction between clusters and cell types. The former is an empirical construct while the latter is a biological truth (albeit a vaguely defined one). For this reason, questions like “what is the true number of clusters?” are usually meaningless. We can define as many clusters as we like, with whatever algorithm we like - each clustering will represent its own partitioning of the high-dimensional expression space, and is as “real” as any other clustering.
It is helpful to realize that clustering, like a microscope, is simply a tool to explore the data. We can zoom in and out by changing the resolution of the clustering parameters, and we can experiment with different clustering algorithms to obtain alternative perspectives of the data.


## With Seurat

```{r seurat}
#only run if first time
if (!file.exists( here("processed", project, "srt_clusters_02.RDS"))) {
  
  srt <- as.Seurat(sce)
  
  # delete old clustering
  srt_cluster_names <- grep("originalexp_snn_res", names(srt@meta.data), value = TRUE)
  srt_clusters_metadata <- srt[[srt_cluster_names]]
  srt_cluster_names <-
    grep("originalexp_snn_res", names(srt_clusters_metadata), value = TRUE)
  for (cluster in srt_cluster_names) {
    colData(sce)[[cluster]] <- NULL
  }
  
  srt <- FindNeighbors(srt, reduction = "harmony", dims = 1:25)
  srt <- FindClusters(srt, 
                      resolution = c(0.01, 0.04, 0.05, 
                                     seq(from = 0.1, to = 1, by = 0.1))
  )
  srt_cluster_names <- grep("originalexp_snn_res", names(srt@meta.data), value = TRUE)
  srt_clusters_metadata <- srt[[srt_cluster_names]]
  
  saveRDS(srt_clusters_metadata, here("processed", project, "srt_clusters_02.RDS"))
  
  plot_list_func(srt, 
                 col_pattern="originalexp_snn_res", 
                 plot_cols = cols,
                 reduction = "TSNE", 
                 label_size = 4 
  )
  
  DimPlot(srt, reduction = "TSNE", group.by = "originalexp_snn_res.1", label = TRUE)
  
} else{
  srt_clusters_metadata <-
    readRDS( here("processed", project, "srt_clusters_02.RDS"))
}

```

```{r fig.height=14, fig.width=14}
clustree(srt_clusters_metadata, prefix = "originalexp_snn_res.", edge_arrow = FALSE)
```

### save clusterings 
```{r save}
if (!file.exists( here("processed", project, "sce_clusters_02.RDS"))) {
# check the cell names are the same
identical(row.names(colData(sce)), row.names(srt_clusters_metadata))

 # delete old clustering from sce
  srt_cluster_names <-
    grep("originalexp_snn_res", names(srt_clusters_metadata), value = TRUE)
  for (cluster in srt_cluster_names) {
  colData(sce)[[cluster]] <- NULL
  }
  
# add them
colData(sce) <- cbind(colData(sce), srt_clusters_metadata)

## save
saveRDS(sce, here("processed", project, "sce_clusters_02.RDS"))
}
```

```{r plot-seurat}
#plot seurat clustering
plot_list_func(sce, 
               col_pattern="originalexp_snn_res", 
               plot_cols = cols,
               reduction = "TSNE", 
               label_size = 0.5 
               )
```
